+++ src/websrv/websrv.service.ts
@@ -41,7 +41,7 @@ export type WebsrvServiceConfig = {
   cert: string;
   key: string;
   ca_bundle: string;
-  proxyTimeout: number;
+  proxyTimeout: number | string;
 }
 
 @Injectable()
@@ -60,7 +60,7 @@ export class WebsrvService {
         target: this._cfg.api_url,
         secure: false,
         ws: true,
-        proxyTimeout: this._cfg.proxyTimeout
+        proxyTimeout: typeof(this._cfg.proxyTimeout) === 'string' ? parseInt(this._cfg.proxyTimeout) : this._cfg.proxyTimeout
       });
     } catch(err) {
       this.log.error(`Error creating proxy server: ${err.message}`);